chore: extract common code to lib
